Recuperacion

+++ Proyecto 2D/Assets/Scripts/Herir.cs
@@ -6,7 +6,7 @@ public class Herir : MonoBehaviour
 {
     // Variables a configurar desde el editor
     [Header("Configuracion")]
-    [SerializeField] float puntos = 5f;
+    [SerializeField] int puntos = 5;
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
